Move all tests for arguments of epyccel to same file. Add test for verbose
